# api-docs.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"카테고리","description":"카테고리 조회 API"},{"name":"메뉴","description":"메뉴 조회·등록·수정·삭제 API"}],"paths":{"/api/menus/{menuCode}":{"get":{"tags":["메뉴"],"summary":"메뉴 상세 조회","description":"result.menu에 조회한 메뉴를 담아 반환한다.","operationId":"findMenuByCode","parameters":[{"name":"menuCode","in":"path","description":"조회할 메뉴 코드","required":true,"schema":{"type":"integer","format":"int32"},"example":1}],"responses":{"200":{"description":"메뉴 상세 조회 성공","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseMessage"}}}},"404":{"description":"메뉴를 찾을 수 없음","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseMessage"}}}},"500":{"description":"서버 내부 오류","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseMessage"}}}}}},"put":{"tags":["메뉴"],"summary":"메뉴 수정","description":"result.menu에 수정된 메뉴를 담아 반환한다.","operationId":"updateMenu","parameters":[{"name":"menuCode","in":"path","description":"수정할 메뉴 코드","required":true,"schema":{"type":"integer","format":"int32"},"example":1}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MenuDTO"}}},"required":true},"responses":{"200":{"description":"메뉴 수정 성공","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseMessage"}}}},"400":{"description":"유효하지 않은 요청 값","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseMessage"}}}},"404":{"description":"메뉴를 찾을 수 없음","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseMessage"}}}},"500":{"description":"서버 내부 오류","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseMessage"}}}}}},"delete":{"tags":["메뉴"],"summary":"메뉴 삭제","description":"실제 HTTP 상태는 200이고 응답 본문의 httpStatus는 204다. result.deletedMenuCode에 삭제한 메뉴 코드를 담아 반환한다.","operationId":"deleteMenu","parameters":[{"name":"menuCode","in":"path","description":"삭제할 메뉴 코드","required":true,"schema":{"type":"integer","format":"int32"},"example":1}],"responses":{"200":{"description":"메뉴 삭제 성공","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseMessage"}}}},"404":{"description":"메뉴를 찾을 수 없음","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseMessage"}}}},"500":{"description":"서버 내부 오류","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseMessage"}}}}}}},"/api/menus":{"get":{"tags":["메뉴"],"summary":"전체 메뉴 조회","description":"result.menus에 전체 메뉴 목록을 담아 반환한다.","operationId":"findAllMenus","responses":{"200":{"description":"메뉴 목록 조회 성공","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseMessage"}}}},"500":{"description":"서버 내부 오류","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseMessage"}}}}}},"post":{"tags":["메뉴"],"summary":"메뉴 등록","description":"menuCode는 서버가 생성한다. result.menu에 등록된 메뉴를 담아 반환한다.","operationId":"saveMenu","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MenuDTO"}}},"required":true},"responses":{"201":{"description":"메뉴 등록 성공","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseMessage"}}}},"400":{"description":"유효하지 않은 요청 값","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseMessage"}}}},"500":{"description":"서버 내부 오류","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseMessage"}}}}}}},"/api/menus/search":{"get":{"tags":["메뉴"],"summary":"가격 기준 메뉴 검색","description":"지정 가격을 초과하는 메뉴를 result.menus에, 기준 가격을 result.searchPrice에 담아 반환한다.","operationId":"findMenusByPrice","parameters":[{"name":"menuPrice","in":"query","description":"초과 조건에 사용할 기준 가격","required":true,"schema":{"type":"integer","format":"int32"},"example":10000}],"responses":{"200":{"description":"가격 기준 메뉴 검색 성공","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseMessage"}}}},"500":{"description":"서버 내부 오류","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseMessage"}}}}}}},"/api/menus/pages":{"get":{"tags":["메뉴"],"summary":"메뉴 페이지 조회","description":"페이지 번호는 1부터 시작한다. result에 content, totalElements, totalPages, size, number, first, last를 담아 반환한다.","operationId":"findMenuPage","parameters":[{"name":"page","in":"query","description":"One-based page index (1..N)","required":false,"schema":{"type":"integer","default":1,"minimum":0}},{"name":"size","in":"query","description":"The size of the page to be returned","required":false,"schema":{"type":"integer","default":10,"minimum":1}},{"name":"sort","in":"query","description":"Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.","required":false,"schema":{"type":"array","items":{"type":"string"}}}],"responses":{"200":{"description":"페이징 메뉴 목록 조회 성공","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseMessage"}}}},"500":{"description":"서버 내부 오류","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseMessage"}}}}}}},"/api/menus/pages/sort":{"get":{"tags":["메뉴"],"summary":"정렬된 메뉴 페이지 조회","description":"페이지 번호는 1부터 시작한다. result에 content, totalElements, totalPages, size, number, first, last, sort, direction을 담아 반환한다.","operationId":"findMenuPageWithSort","parameters":[{"name":"page","in":"query","description":"One-based page index (1..N)","required":false,"schema":{"type":"integer","default":1,"minimum":0}},{"name":"size","in":"query","description":"The size of the page to be returned","required":false,"schema":{"type":"integer","default":5,"minimum":1}},{"name":"sort","in":"query","description":"Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.","required":false,"schema":{"type":"array","items":{"type":"string"}}},{"name":"sortBy","in":"query","description":"정렬 기준 필드","required":false,"schema":{"type":"string","default":"menuPrice"},"example":"menuPrice"},{"name":"direction","in":"query","description":"정렬 방향","required":false,"schema":{"type":"string","default":"asc"},"example":"asc"}],"responses":{"200":{"description":"정렬된 페이징 메뉴 목록 조회 성공","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseMessage"}}}},"500":{"description":"서버 내부 오류","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseMessage"}}}}}}},"/api/categories":{"get":{"tags":["카테고리"],"summary":"전체 카테고리 조회","description":"result.categories에 전체 카테고리 목록을 담아 반환한다.","operationId":"findAllCategories","responses":{"200":{"description":"카테고리 목록 조회 성공","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseMessage"}}}},"500":{"description":"서버 내부 오류","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseMessage"}}}}}}},"/api/categories/{categoryCode}":{"get":{"tags":["카테고리"],"summary":"카테고리 상세 조회","description":"result.category에 조회한 카테고리를 담아 반환한다.","operationId":"findCategoryByCode","parameters":[{"name":"categoryCode","in":"path","description":"조회할 카테고리 코드","required":true,"schema":{"type":"integer","format":"int32"},"example":1}],"responses":{"200":{"description":"카테고리 상세 조회 성공","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseMessage"}}}},"404":{"description":"카테고리를 찾을 수 없음","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseMessage"}}}},"500":{"description":"서버 내부 오류","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseMessage"}}}}}}}},"components":{"schemas":{"MenuDTO":{"type":"object","properties":{"menuCode":{"type":"integer","format":"int32"},"menuName":{"type":"string"},"menuPrice":{"type":"integer","format":"int32"},"categoryCode":{"type":"integer","format":"int32"},"categoryName":{"type":"string"},"orderableStatus":{"type":"string"}}},"ResponseMessage":{"type":"object","properties":{"httpStatus":{"type":"integer","format":"int32"},"message":{"type":"string"},"result":{"type":"object","additionalProperties":{}}}}}}}